Complete task 0 and 1 of the printf project
The _printf function was revised and updated to make it more SOLID. In
this new version, more functionaly can be added to the function without
violating any betty regulations. Previously, this would not have been
possible because the function was utilising a switch statement.
The functions that handle the format specifiers were transfered to their
individual source files.