Complete task 2, 3 and 4 of the printf project
New files were added to take care of the aforementioned tasks. A new
feature was also added where all characters to be printed are instead
added to a buffer of size 1024 bytes to reduce the number of times
write() is called.
u_o_x_handlers.c contains functions for the u, o x, and X format
specifiers.
s_handler.c contains the function for the custom S specifier.